requirements_lightning.txt
# Core PyTorch Lightning and ML libraries
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
lightning>=2.0.0

# Transformers and Diffusers
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Optimization libraries
bitsandbytes>=0.41.0
deepspeed>=0.9.0

# Data handling
datasets>=2.12.0
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0  # Optional, for W&B logging
tqdm>=4.65.0

# Image processing
opencv-python>=4.7.0  # Optional, for advanced image processing
albumentations>=1.3.0  # Optional, for data augmentation

# Development tools (optional)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 